hide portal features on viewer page
